Refactor OpalDefendantAccountHistoryService to utilize AmendmentHistorySourceService for fetching amendments
